+++ ncmc_lib/src/lib.rs
@@ -10,7 +10,8 @@ use id3::{
     frame::{Picture, PictureType},
     Tag, TagLike as _,
 };
-use serde::{Deserialize, Serialize};
+use serde::{ser::SerializeTuple as _, Deserialize, Serialize};
+use serde_json::Value;
 use std::{
     fs::File,
     io::{Read, Seek, Write as _},
@@ -39,7 +40,7 @@ impl NcmFile {
         Self::verify_header(&mut file)?;
         let key = Key::new(Self::get_key(&mut file)?);
         let mut meta = Self::get_meta(&mut file)?;
-        file.seek_relative(9)?; // CRC(4) and Padding(5)
+        file.seek_relative(5)?; // CRC(5)
         meta.cover = Self::get_cover(&mut file)?;
         Ok(Self {
             file,
@@ -47,6 +48,30 @@ impl NcmFile {
             key,
             meta,
         })
+    }
+
+    /// If cover in meta is empty, download it from the meta.album_pic
+    #[cfg(feature = "cover_download")]
+    pub fn with_cover(mut self) -> Result<Self> {
+        if self.meta.cover.is_empty() {
+            self.meta.cover = ureq::get(&self.meta.album_pic)
+                .call()?
+                .body_mut()
+                .read_to_vec()?;
+        }
+        Ok(self)
+    }
+
+    /// `with_cover` but in place
+    #[cfg(feature = "cover_download")]
+    pub fn fetch_cover(&mut self) -> Result<()> {
+        if self.meta.cover.is_empty() {
+            self.meta.cover = ureq::get(&self.meta.album_pic)
+                .call()?
+                .body_mut()
+                .read_to_vec()?;
+        }
+        Ok(())
     }
 
     fn verify_header(file: &mut File) -> Result<()> {
@@ -117,9 +142,12 @@ impl NcmFile {
     fn get_cover(file: &mut File) -> Result<Vec<u8>> {
         let mut buf = [0; 4];
         file.read_exact(&mut buf)?;
+        let cover_frame_length = u32::from_le_bytes(buf);
+        file.read_exact(&mut buf)?;
         let length = u32::from_le_bytes(buf);
         let mut buf = vec![0; length as usize];
         file.read_exact(&mut buf)?;
+        file.seek_relative((cover_frame_length - length) as i64)?;
         Ok(buf)
     }
 
@@ -132,19 +160,19 @@ impl NcmFile {
     /// save as general format to the specified path
     pub fn save_to(self, path: impl AsRef<Path>) -> Result<()> {
         let tag = Tag::from(&self.meta);
-        self.save_without_tags_to(&path)?;
+        self.save_without_meta_to(&path)?;
         tag.write_to_path(path, id3::Version::Id3v24)?;
         Ok(())
     }
 
     /// savenext to the original ncm file without tags
-    pub fn save_without_tags(self) -> Result<()> {
+    pub fn save_without_meta(self) -> Result<()> {
         let path = self.path.with_extension(&self.meta.format);
-        self.save_without_tags_to(path)
+        self.save_without_meta_to(path)
     }
 
     /// save to the specified path without tags
-    pub fn save_without_tags_to(mut self, path: impl AsRef<Path>) -> Result<()> {
+    pub fn save_without_meta_to(mut self, path: impl AsRef<Path>) -> Result<()> {
         let mut file = std::fs::File::create(&path)?;
         std::io::copy(&mut self, &mut file)?;
         file.flush()?;
@@ -196,30 +224,90 @@ impl Iterator for Key {
 #[allow(missing_docs)]
 pub struct Meta {
     pub album: String,
-    #[serde(rename = "albumId")]
-    pub album_id: usize,
+    #[serde(rename = "albumId", deserialize_with = "deserialize_to_string")]
+    pub album_id: String,
     /// The url of the cover image
     #[serde(rename = "albumPic")]
     pub album_pic: String,
-    #[serde(rename = "albumPicDocId")]
-    pub album_pic_doc_id: serde_json::Value,
+    #[serde(rename = "albumPicDocId", deserialize_with = "deserialize_to_string")]
+    pub album_pic_doc_id: String,
     pub alias: Vec<String>,
-    pub artist: Vec<(String, usize)>,
+    pub artist: Vec<Artist>,
     pub bitrate: usize,
     pub duration: usize,
+    pub fee: Option<usize>,
     pub flag: Option<usize>,
     pub format: String,
+    #[serde(rename = "mp3DocId")]
+    pub mp3_doc_id: Option<String>,
     pub gain: Option<f64>,
-    #[serde(rename = "musicId")]
-    pub music_id: usize,
+    #[serde(rename = "musicId", deserialize_with = "deserialize_to_string")]
+    pub music_id: String,
     #[serde(rename = "musicName")]
     pub music_name: String,
-    #[serde(rename = "mvId")]
-    pub mv_id: usize,
+    #[serde(rename = "mvId", deserialize_with = "deserialize_to_string")]
+    pub mv_id: String,
     #[serde(rename = "transNames")]
     pub trans_names: Vec<String>,
     #[serde(skip)]
     pub cover: Vec<u8>,
+}
+
+#[derive(Debug, Clone)]
+pub struct Artist {
+    name: String,
+    id: String,
+}
+
+impl Serialize for Artist {
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let mut seq = serializer.serialize_tuple(2)?;
+        seq.serialize_element(&self.name)?;
+        seq.serialize_element(&self.id)?;
+        seq.end()
+    }
+}
+
+impl<'de> Deserialize<'de> for Artist {
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let v = Value::deserialize(deserializer)?;
+        match v.as_array() {
+            Some(v) => {
+                if let [Value::String(name), Value::String(id)] = v.as_slice() {
+                    Ok(Artist {
+                        name: name.clone(),
+                        id: id.clone(),
+                    })
+                } else if let [Value::String(name), Value::Number(id)] = v.as_slice() {
+                    Ok(Artist {
+                        name: name.clone(),
+                        id: id.to_string(),
+                    })
+                } else {
+                    Err(serde::de::Error::custom("Invalid value"))
+                }
+            }
+            None => Err(serde::de::Error::custom("Invalid value")),
+        }
+    }
+}
+
+fn deserialize_to_string<'de, D>(deserializer: D) -> core::result::Result<String, D::Error>
+where
+    D: serde::Deserializer<'de>,
+{
+    let s = Value::deserialize(deserializer)?;
+    match s {
+        Value::String(s) => Ok(s),
+        Value::Number(n) => Ok(n.to_string()),
+        _ => Err(serde::de::Error::custom("Invalid value")),
+    }
 }
 
 impl From<&Meta> for Tag {
@@ -232,18 +320,16 @@ impl From<&Meta> for Tag {
             description: "Cover".to_string(),
             data: meta.cover.clone(),
         });
-        tag.set_artist(
-            meta.artist
-                .iter()
-                .map(|(name, _)| name)
-                .fold(String::new(), |acc, x| {
-                    if acc.is_empty() {
-                        x.to_string()
-                    } else {
-                        acc + ", " + x
-                    }
-                }),
-        );
+        tag.set_artist(meta.artist.iter().map(|artist| &artist.name).fold(
+            String::new(),
+            |acc, x| {
+                if acc.is_empty() {
+                    x.to_string()
+                } else {
+                    acc + ", " + x
+                }
+            },
+        ));
         tag.set_duration(meta.duration as u32);
         tag.set_title(meta.music_name.clone());
         tag